backend/requirements.txt
fastapi==0.111.0
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.4
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.1
dateparser==1.2.0
langchain==0.3.26
langchain-google-genai==2.0.10
google-generativeai==0.8.0
bcrypt==3.2.2